# src/swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Gym App API",
    "version": "1.0.0",
    "description": "API for Gym App"
  },
  "host": "localhost:3000",
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/mesocycle/users/{uuid}": {
      "get": {
        "tags": [
          "Mesocycle"
        ],
        "summary": "Get a mesocycle by user ID",
        "description": "Retrieves a mesocycle by user ID.",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Mesocycles found."
          },
          "404": {
            "description": "Mesocycles not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/mesocycle/all": {
      "get": {
        "tags": [
          "Mesocycle"
        ],
        "summary": "Get all mesocycles",
        "description": "Retrieves all mesocycles.",
        "responses": {
          "200": {
            "description": "Mesocycles found."
          },
          "404": {
            "description": "Mesocycles not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/mesocycle/{uuid}": {
      "get": {
        "tags": [
          "Mesocycle"
        ],
        "summary": "Get a mesocycle by ID",
        "description": "Retrieves a mesocycle by its ID.",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Mesocycle ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Mesocycle found."
          },
          "404": {
            "description": "Mesocycle not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/mesocycle/name/{name}": {
      "get": {
        "tags": [
          "Mesocycle"
        ],
        "summary": "Get a mesocycle by name",
        "description": "Retrieves a mesocycle by its name.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Mesocycle name"
          }
        ],
        "responses": {
          "200": {
            "description": "Mesocycle found."
          },
          "404": {
            "description": "Mesocycle not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/mesocycle/{uuid}/workout": {
      "get": {
        "tags": [
          "Mesocycle"
        ],
        "summary": "Get workouts by mesocycle ID",
        "description": "Retrieves workouts by mesocycle ID.",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Mesocycle ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Workouts found."
          },
          "404": {
            "description": "Workouts not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/mesocycle/{uuid}/users": {
      "get": {
        "tags": [
          "Mesocycle"
        ],
        "summary": "Get users by mesocycle ID",
        "description": "Retrieves users by mesocycle ID.",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Mesocycle ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Users found."
          },
          "404": {
            "description": "Users not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
...
    "/mesocycle/create/{uuid}": {
      "post": {
        "tags": [
          "Mesocycle"
        ],
        "summary": "Create a new mesocycle",
        "description": "Creates a new mesocycle.",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "User ID"
          },
          {
            "name": "name",
            "description": "Mesocycle name",
            "in": "query",
            "type": "string"
          },
          {
            "name": "length",
            "description": "Mesocycle length",
            "in": "query",
            "type": "string"
          },
          {
            "name": "phase",
            "description": "Mesocycle phase",
            "in": "query",
            "type": "string"
          },
          {
            "name": "periodization",
            "description": "Mesocycle periodization",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Mesocycle created."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/mesocycle/update/{uuid}": {
      "put": {
        "tags": [
          "Mesocycle"
        ],
        "summary": "Update a mesocycle by ID",
        "description": "Updates a mesocycle.",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Mesocycle ID"
          },
          {
            "name": "name",
            "description": "Mesocycle name",
            "in": "query",
            "type": "string"
          },
          {
            "name": "length",
            "description": "Mesocycle length",
            "in": "query",
            "type": "string"
          },
          {
            "name": "phase",
            "description": "Mesocycle phase",
            "in": "query",
            "type": "string"
          },
          {
            "name": "periodization",
            "description": "Mesocycle periodization",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Mesocycle updated."
          },
          "404": {
            "description": "Mesocycle not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
...
    "/users/all": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get all users",
        "description": "Retrieves all users.",
        "responses": {
          "200": {
            "description": "Users found."
          },
          "404": {
            "description": "Users not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/users/{uuid}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get a user by ID",
        "description": "Retrieves a user by their ID.",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User found."
          },
          "404": {
            "description": "User not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/users/email/{email}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get a user by email",
        "description": "Retrieves a user by their email.",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "User email"
          }
        ],
        "responses": {
          "200": {
            "description": "User found."
          },
          "404": {
            "description": "User not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/users/username/{username}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get a user by username",
        "description": "Retrieves a user by their username.",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "User username"
          }
        ],
        "responses": {
          "200": {
            "description": "User found."
          },
          "404": {
            "description": "User not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
...
  }
}